@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
@@ -25,11 +26,11 @@ namespace ModCore.CoreApi
 		// This method gets called by the runtime. Use this method to add services to the container.
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.AddMvc();
+			services.AddControllersWithViews();
 		}
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
 		{
 			if (env.IsDevelopment())
 			{
@@ -50,7 +51,10 @@ namespace ModCore.CoreApi
 			app.UseDefaultFiles(df);
 			app.UseStaticFiles(new StaticFileOptions() { FileProvider = fp, RequestPath = "" });
 
-			app.UseMvc();
+			app.UseRouting();
+			app.UseEndpoints(x => {
+				x.MapControllers();
+			});
 		}
 	}
 }
